requirements.txt
astroid==1.5.2
isort==4.2.5
lazy-object-proxy==1.3.1
mccabe==0.6.1
pylint==1.7.1
PyYAML==3.12
six==1.10.0
wrapt==1.10.10
